+++ xolotl/xolotlCore/advection/SurfaceAdvectionHandler.cpp
@@ -129,7 +129,7 @@ void SurfaceAdvectionHandler::computeAdvection(const IReactionNetwork& network,
 }
 
 void SurfaceAdvectionHandler::computePartialsForAdvection(
-		const IReactionNetwork& network, double *val, int *indices,
+		const IReactionNetwork& network, double *val, long int *indices,
 		const NDPoint<3>& pos, double hxLeft, double hxRight, int ix, double hy,
 		int iy, double hz, int iz) const {
 	// Consider each advecting cluster.
