+++ xolotl/xolotlCore/diffusion/IDiffusionHandler.h
@@ -95,7 +95,7 @@ public:
 	 * @param iz The position on the z grid
 	 */
 	virtual void computePartialsForDiffusion(const IReactionNetwork& network,
-			double *val, int *indices, double hxLeft, double hxRight, int ix,
+			double *val, long int *indices, double hxLeft, double hxRight, int ix,
 			double sy = 0.0, int iy = 0, double sz = 0.0,
 			int iz = 0) const = 0;
