+++ xolotl/xolotlCore/modifiedreaction/trapmutation/ITrapMutationHandler.h
@@ -155,7 +155,7 @@ public:
 	 * at this grid point
 	 */
 	virtual int computePartialsForTrapMutation(const IReactionNetwork& network,
-			double *val, int *indices, int xi, int yj = 0, int zk = 0) = 0;
+			double *val, long int *indices, int xi, int yj = 0, int zk = 0) = 0;
 
 	/**
 	 * Get the total number of clusters in the network that can undergo trap mutation.
