+++ xolotl/xolotlCore/reactants/feclusters/FeClusterReactionNetwork.cpp
@@ -879,7 +879,7 @@ void FeClusterReactionNetwork::computeAllFluxes(double *updatedConcOffset,
 }
 
 void FeClusterReactionNetwork::computeAllPartials(
-		const std::vector<size_t>& startingIdx, const std::vector<int>& indices,
+		const std::vector<size_t>& startingIdx, const std::vector<long int>& indices,
 		std::vector<double>& vals, int i) const {
 	// Initial declarations
 	const int dof = getDOF();
